Fix TRAVIS

+++ pyLIMA/microlsimulator.py
@@ -90,8 +90,8 @@ def time_simulation(time_start, time_end, sampling, bad_weather_percentage):
 
     total_number_of_days = int(time_end - time_start)
     time_step_observations = sampling / 24.0
-    number_of_day_exposure = np.floor(
-        1.0 / time_step_observations)  # less than expected total, more likely in a telescope :)
+    number_of_day_exposure = int(np.floor(
+        1.0 / time_step_observations))  # less than expected total, more likely in a telescope :)
     night_begin = time_start
 
     time_observed = []
@@ -101,7 +101,7 @@ def time_simulation(time_start, time_end, sampling, bad_weather_percentage):
 
         if good_weather > bad_weather_percentage:
             random_begin_of_the_night = 0
-            night_end = night_begin + 1.0
+            night_end = night_begin + 1
             time_observed += np.linspace(night_begin + time_step_observations + random_begin_of_the_night, night_end,
                                          number_of_day_exposure).tolist()
